+++ R/IBDprob.R
@@ -112,12 +112,18 @@ c.IBDprob <- function(...) {
 
 #' Plot function for objects of class IBDprob
 #'
-#' Creates a plot for an object of class \code{IBDprob}.
+#' Creates a plot for an object of class \code{IBDprob}. Two types of plot can
+#' be made, a plot for a single genotype showing the IBD probabilities for all
+#' parents across the genome (\code{plotType = "singleGeno"}), or a plot showing
+#' for all genotypes the probabilities of the parent with the highest
+#' probability per marker (\code{plotType = "allGeno"}).
 #'
 #' @param x An object of class \code{IBDprob}.
 #' @param ... Further arguments. Unused.
+#' @param plotType A character string indicating the type of plot that should
+#' be made.
 #' @param genotype A character string indicating the genotype for which the
-#' plot should be made.
+#' plot should be made. Ignored if \code{plotType = "allGeno"}.
 #' @param title A character string, the title of the plot.
 #' @param output Should the plot be output to the current device? If
 #' \code{FALSE}, only a ggplot object is invisibly returned.
@@ -137,46 +143,94 @@ c.IBDprob <- function(...) {
 #'
 #' ## Plot results for genotype dh005.
 #' plot(SxMIBD_Ext,
+#'      plotType = "singleGeno",
 #'      genotype = "dh005")
+#'
+#' ## Plot results for all genotypes.
+#' plot(SxMIBD_Ext,
+#'      plotType = "allGeno")
 #'
 #' @export
 plot.IBDprob <- function(x,
                          ...,
+                         plotType = c("singleGeno", "allGeno"),
                          genotype,
-                         title = genotype,
+                         title = NULL,
                          output = TRUE) {
   map <- x$map
   markers <- x$markers
+  parents <- x$parents
   ## Input checks.
-  if (!inherits(genotype, "character") || length(genotype) > 1) {
-    stop("genotype should be a character string.\n")
-  }
-  if (!inherits(title, "character") || length(title) > 1) {
+  plotType <- match.arg(plotType)
+  if (!is.null(title) && (!inherits(title, "character") || length(title) > 1)) {
     stop("title should be a character string.\n")
   }
-  if (!genotype %in% dimnames(markers)[[2]]) {
-    stop(paste("genotype",genotype,"not defined\n"))
+  if (plotType == "singleGeno") {
+    if (!inherits(genotype, "character") || length(genotype) > 1) {
+      stop("genotype should be a character string.\n")
+    }
+    if (!genotype %in% dimnames(markers)[[2]]) {
+      stop(paste("genotype", genotype, "not defined\n"))
+    }
+    ## Convert to long format for plotting.
+    markersLong <- markers3DtoLong(x)
+    ## Merge map info to probabilities.
+    plotDat <- merge(markersLong, map, by.x = "snp", by.y = "row.names")
+    ## Restrict to selected genotype.
+    plotDat <- plotDat[plotDat[["genotype"]] == genotype, ]
+    ## Construct title.
+    if (is.null(title)) {
+      title <- genotype
+    }
+    p <- ggplot2::ggplot(plotDat,
+                         ggplot2::aes_string(x = "pos", y = "parent",
+                                             fill = "prob")) +
+      ggplot2::geom_tile(width = 3) +
+      ggplot2::facet_grid(". ~ chr", scales = "free", space = "free",
+                          switch = "both") +
+      ggplot2::scale_fill_gradient(low = "white", high = "black") +
+      ggplot2::scale_x_continuous(expand = c(0, 0)) +
+      ggplot2::scale_y_discrete(expand = c(0, 0)) +
+      ggplot2::labs(title = title) +
+      ggplot2::theme(
+        plot.title = ggplot2::element_text(hjust = 0.5),
+        panel.background = ggplot2::element_blank(),
+        panel.spacing.x = ggplot2::unit(4, "mm"))
+  } else if (plotType == "allGeno") {
+    ## Get max IBD value and parent per marker-position combination.
+    maxVals <- apply(X = markers, MARGIN = 1:2, FUN = max)
+    maxPars <- parents[apply(X = markers, MARGIN = 1:2, FUN = which.max)]
+    nGeno <- nrow(maxVals)
+    ## Create plot data.
+    plotDat <- data.frame(marker = factor(dimnames(maxVals)[[1]],
+                                          levels = dimnames(maxVals)[[1]]),
+                          genotype = rep(dimnames(maxVals)[[2]], each = nGeno),
+                          maxVal = as.vector(maxVals),
+                          maxPar = as.vector(maxPars))
+    ## Construct title.
+    if (is.null(title)) {
+      title <- "IBD probabilities across the genome for all genotypes"
+    }
+    ## Get positions of start of new chromosomes.
+    newChrs <- which(!duplicated(map[["chr"]]))[-1] - 0.5
+    p <- ggplot2::ggplot(data = plotDat,
+                         ggplot2::aes_string(x = "marker", y = "genotype",
+                                             alpha = "maxVal",
+                                             fill = "maxPar"))+
+      ggplot2::geom_raster() +
+      ggplot2::labs(title = title, x = "Genome", y = "Genotypes",
+                    fill = "Parent", alpha = "Probability") +
+      ggplot2::geom_vline(xintercept = newChrs, linetype = "dashed",
+                          color = "black") +
+      ggplot2::theme(
+        panel.background = ggplot2::element_blank(),
+        plot.background = ggplot2::element_blank(),
+        axis.text = ggplot2::element_blank(),
+        axis.ticks = ggplot2::element_blank(),
+        panel.border = ggplot2::element_rect(fill = NA),
+        plot.title = ggplot2::element_text(hjust = 0.5)
+      )
   }
-  ## Convert to long format for plotting.
-  markersLong <- markers3DtoLong(x)
-  ## Merge map info to probabilities.
-  plotDat <- merge(markersLong, map, by.x = "snp", by.y = "row.names")
-  ## Restrict to selected genotype.
-  plotDat <- plotDat[plotDat[["genotype"]] == genotype, ]
-  p <- ggplot2::ggplot(plotDat,
-                  ggplot2::aes_string(x = "pos", y = "parent",
-                                      fill = "prob")) +
-    ggplot2::geom_tile(width = 3) +
-    ggplot2::facet_grid(". ~ chr", scales = "free", space = "free",
-                        switch = "both") +
-    ggplot2::scale_fill_gradient(low = "white", high = "black") +
-    ggplot2::scale_x_continuous(expand = c(0, 0)) +
-    ggplot2::scale_y_discrete(expand = c(0, 0)) +
-    ggplot2::labs(title = title) +
-    ggplot2::theme(
-      plot.title = ggplot2::element_text(hjust = 0.5),
-      panel.background = ggplot2::element_blank(),
-      panel.spacing.x = ggplot2::unit(4, "mm"))
   if (output) {
     plot(p)
   }