+++ src/pedigree/read_map.h
@@ -18,13 +18,9 @@ LinkageMap read_map_file(const std::string& mapfile);
 
 void print_marker_warnings(const LinkageMap& markermap, const std::vector<std::string>& markers);
 
-LinkageMap read_eval_pos_file(const std::string& filename);
-
 LinkageMap read_eval_pos_df(const Rcpp::DataFrame& evalposdf);
 
-LinkageMap select_chr(const LinkageMap& markermap, int sel_chr);
+LinkageMap select_chr(const LinkageMap& markermap, std::string sel_chr);
 
 #endif
-
-